% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_grid}
\alias{control_grid}
\alias{new_backend_options}
\title{Control aspects of the grid search process}
\usage{
control_grid(
  verbose = FALSE,
  allow_par = TRUE,
  extract = NULL,
  save_pred = FALSE,
  pkgs = NULL,
  save_workflow = FALSE,
  event_level = "first",
  parallel_over = NULL,
  backend_options = NULL
)

new_backend_options(..., class = character())
}
\description{
Control aspects of the grid search process
}
\keyword{internal}
