% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/utils.R
\name{check_rset}
\alias{check_rset}
\alias{check_parameters}
\alias{check_workflow}
\alias{check_metrics}
\alias{check_initial}
\alias{val_class_or_null}
\alias{val_class_and_single}
\alias{empty_ellipses}
\alias{is_recipe}
\alias{is_preprocessor}
\alias{is_workflow}
\title{Internal functions for developers}
\usage{
check_rset(x)

check_parameters(wflow, pset = NULL, data, grid_names = character(0))

check_workflow(x, ..., pset = NULL, check_dials = FALSE, call = caller_env())

check_metrics(x, object)

check_initial(
  x,
  pset,
  wflow,
  resamples,
  metrics,
  eval_time,
  ctrl,
  checks = "grid"
)

val_class_or_null(x, cls = "numeric", where = NULL)

val_class_and_single(x, cls = "numeric", where = NULL)

empty_ellipses(...)

is_recipe(x)

is_preprocessor(x)

is_workflow(x)
}
\arguments{
\item{x}{An object.}

\item{wflow}{A \code{workflow} object.}

\item{data}{The training data.}

\item{grid_names}{A character vector of column names from the grid.}

\item{...}{Other options}

\item{check_dials}{A logical for check for a NULL parameter object.}

\item{object}{A \code{workflow} object.}

\item{resamples}{An \code{rset} object.}

\item{ctrl}{A \code{control_grid} object.}

\item{cls}{A character vector of possible classes}

\item{where}{A character string for the calling function.}
}
\description{
These are not intended for use by the general public.
}
\keyword{internal}
