% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopy.R
\name{loopy}
\alias{loopy}
\title{Iterate over workflow settings}
\usage{
loopy(split, grid, static)
}
\arguments{
\item{grid}{A parameter grid.}

\item{sched}{A preprocessing schedule.}
}
\description{
Iterate over workflow settings
}
